template component with good click behaviour
